+++ src/main/java/com/nhom7/coworkingspace/util/ValidImage.java
@@ -22,4 +22,3 @@ public @interface ValidImage {
 
     long maxSizeInBytes() default 5 * 1024 * 1024; // 5MB
 }
-